requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
python-dotenv==1.0.1
Werkzeug==3.0.1
email-validator==2.1.0.post1
requests>=2.31.0
bcchapi>=1.1.2
pandas>=1.2.0
transbank-sdk>=1.0.0
marshmallow>=3.0.0 